fix test coverage and make sure file exists before trying to open it

+++ version.py
@@ -213,11 +213,12 @@ async def update_version_file(*, new_version, working_dir, readonly):
     """
     old_version = None
     version_file = Path(working_dir) / "VERSION"
-    with open(version_file, "r") as f:
-        old_version = f.readline()
-    if not readonly:
-        with open(version_file, "w") as f:
-            f.write(new_version)
+    if version_file.is_file():
+        with open(version_file, "r") as f:
+            old_version = f.readline()
+        if not readonly:
+            with open(version_file, "w") as f:
+                f.write(new_version)
     return old_version
